Intermediate, partly working state after blaze rework.

+++ blazert/bvh/statistics.h
@@ -9,14 +9,14 @@ public:
   unsigned int max_tree_depth;
   unsigned int num_leaf_nodes;
   unsigned int num_branch_nodes;
-  float build_secs;
+  double build_secs;
 
-  // TODO: This seems to be not done: Set default value: Taabb = 0.2
   BVHBuildStatistics()
       : max_tree_depth(0),
         num_leaf_nodes(0),
         num_branch_nodes(0),
-        build_secs(0.0f) {}
+        build_secs(0.0) {}
 };
 }// namespace blazert
+
 #endif// BLAZERT_BVH_STATISTICS_H
